backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
inflection==0.5.1
PyJWT==2.7.0
python-dotenv==0.21.0
sqlparse==0.4.3
tzdata==2022.6
async-timeout==4.0.2
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
pyOpenSSL==23.1.1
pytz==2023.3
redis==4.5.4
sqlparse==0.4.3

